Added --features flag to the run command
To enable specific features on the target when executing `cargo ziggy
run`, the -F / --features flag can now be used. It is simply passed
along to rustc.

This can be useful when examining a crash in a more complex target. For
example, verbose logging in the target could be hidden behind a
"logging" feature, to reduce noise when fuzzing. In case the output is
needed for crash analysis, it could simply be activated by running
`cargo ziggy run -F logging`.